test: remove importer dependency from dc solver and add it as fixture
