+++ packages/dc_solver_pkg/src/toop_engine_dc_solver/example_grids.py
@@ -25,7 +25,6 @@ import pypowsybl
 from beartype.typing import Literal, Optional
 from fsspec.implementations.dirfs import DirFileSystem
 from networkx.algorithms.community import kernighan_lin_bisection
-from toop_engine_dc_solver.preprocess.convert_to_jax import load_grid
 from toop_engine_dc_solver.preprocess.pandapower.pandapower_backend import PandaPowerBackend
 from toop_engine_dc_solver.preprocess.powsybl.powsybl_backend import PowsyblBackend
 from toop_engine_grid_helpers.pandapower.example_grids import (
@@ -37,14 +36,12 @@ from toop_engine_grid_helpers.pandapower.example_grids import (
 from toop_engine_grid_helpers.pandapower.pandapower_id_helpers import SEPARATOR
 from toop_engine_grid_helpers.powsybl.example_grids import (
     create_busbar_b_in_ieee,
-    create_complex_grid_battery_hvdc_svc_3w_trafo,
     extract_station_info_powsybl,
     powsybl_case30_with_psts,
     powsybl_case9241,
     powsybl_extended_case57,
 )
 from toop_engine_grid_helpers.powsybl.loadflow_parameters import DISTRIBUTED_SLACK
-from toop_engine_importer.pypowsybl_import import preprocessing
 from toop_engine_interfaces.asset_topology import (
     Busbar,
     BusbarCoupler,
@@ -56,12 +53,6 @@ from toop_engine_interfaces.backend import BackendInterface
 from toop_engine_interfaces.folder_structure import (
     NETWORK_MASK_NAMES,
     PREPROCESSING_PATHS,
-)
-from toop_engine_interfaces.messages.preprocess.preprocess_commands import (
-    AreaSettings,
-    CgmesImporterParameters,
-    PreprocessParameters,
-    UcteImporterParameters,
 )
 
 
@@ -995,90 +986,3 @@ def node_breaker_folder_powsybl(folder: Path) -> None:
     """Copy over all data from the data folder"""
     source = Path(__file__).parent.parent.parent / "tests" / "files" / "test_grid_node_breaker"
     shutil.copytree(source, folder, dirs_exist_ok=True)
-
-
-def create_complex_grid_battery_hvdc_svc_3w_trafo_data_folder(folder: Path) -> None:
-    """Create a preprocessed folder for create_complex_grid_battery_hvdc_svc_3w_trafo().
-
-    Runs the importer and preprocessing.
-
-    Parameter:
-    folder: Path
-        The root folder where the data is saved to.
-    """
-    net = create_complex_grid_battery_hvdc_svc_3w_trafo()
-    pypowsybl.loadflow.run_dc(net, DISTRIBUTED_SLACK)
-
-    output_path_grid = folder / PREPROCESSING_PATHS["grid_file_path_powsybl"]
-    output_path_grid.parent.mkdir(parents=True, exist_ok=True)
-    net.save(output_path_grid)
-    output_path_masks = folder / PREPROCESSING_PATHS["masks_path"]
-    output_path_masks.mkdir(parents=True, exist_ok=True)
-
-    importer_parameters = CgmesImporterParameters(
-        grid_model_file=output_path_grid,
-        data_folder=folder,
-        area_settings=AreaSettings(
-            cutoff_voltage=1,
-            control_area=[""],
-            view_area=[""],
-            nminus1_area=[""],
-            cross_border_limits_n0=None,
-            cross_border_limits_n1=None,
-        ),
-    )
-    preprocessing_parameters = PreprocessParameters(action_set_clip=2**4, enable_bb_outage=False, bb_outage_as_nminus1=False)
-
-    _import_result = preprocessing.convert_file(importer_parameters=importer_parameters)
-    filesystem_dir = DirFileSystem(str(folder))
-    _info, _static_information, _ = load_grid(
-        data_folder_dirfs=filesystem_dir,
-        pandapower=False,
-        status_update_fn=None,
-        parameters=preprocessing_parameters,
-    )
-
-
-def create_ucte_data_folder(folder: Path, ucte_file: Path) -> None:
-    """Create a preprocessed folder for an ucte file.
-
-    Runs the importer and preprocessing.
-
-    Parameter:
-    folder: Path
-        The root folder where the data is saved to.
-    ucte_file: Path
-        The path to the UCTE file to load.
-    """
-    net = pypowsybl.network.load(ucte_file)
-    pypowsybl.loadflow.run_dc(net, DISTRIBUTED_SLACK)
-
-    output_path_grid = folder / PREPROCESSING_PATHS["grid_file_path_powsybl"]
-    output_path_grid.parent.mkdir(parents=True, exist_ok=True)
-    net.save(output_path_grid)
-    output_path_masks = folder / PREPROCESSING_PATHS["masks_path"]
-    output_path_masks.mkdir(parents=True, exist_ok=True)
-
-    importer_parameters = UcteImporterParameters(
-        grid_model_file=output_path_grid,
-        data_folder=folder,
-        area_settings=AreaSettings(
-            cutoff_voltage=1,
-            control_area=[""],
-            view_area=[""],
-            nminus1_area=[""],
-            cross_border_limits_n0=None,
-            cross_border_limits_n1=None,
-        ),
-    )
-    preprocessing_parameters = PreprocessParameters(action_set_clip=2**4, enable_bb_outage=False, bb_outage_as_nminus1=False)
-
-    _import_result = preprocessing.convert_file(importer_parameters=importer_parameters)
-
-    filesystem_dir = DirFileSystem(str(folder))
-    _info, _static_information, _ = load_grid(
-        data_folder_dirfs=filesystem_dir,
-        pandapower=False,
-        status_update_fn=None,
-        parameters=preprocessing_parameters,
-    )
