+++ packages/dc_solver_pkg/tests/conftest.py
@@ -41,8 +41,6 @@ from toop_engine_dc_solver.example_grids import (
     case14_pandapower,
     case30_with_psts,
     case57_data_powsybl,
-    create_complex_grid_battery_hvdc_svc_3w_trafo_data_folder,
-    create_ucte_data_folder,
     node_breaker_folder_powsybl,
     oberrhein_data,
 )
@@ -89,7 +87,10 @@ from toop_engine_grid_helpers.pandapower.pandapower_id_helpers import table_id
 from toop_engine_grid_helpers.powsybl.example_grids import (
     basic_node_breaker_network_powsybl,
     case14_matching_asset_topo_powsybl,
+    create_complex_grid_battery_hvdc_svc_3w_trafo,
 )
+from toop_engine_grid_helpers.powsybl.loadflow_parameters import DISTRIBUTED_SLACK
+from toop_engine_importer.pypowsybl_import import preprocessing
 from toop_engine_interfaces.asset_topology import (
     Busbar,
     BusbarCoupler,
@@ -105,7 +106,10 @@ from toop_engine_interfaces.folder_structure import (
     PREPROCESSING_PATHS,
 )
 from toop_engine_interfaces.messages.preprocess.preprocess_commands import (
+    AreaSettings,
+    CgmesImporterParameters,
     PreprocessParameters,
+    UcteImporterParameters,
 )
 
 chex.set_n_cpu_devices(2)
@@ -1163,4 +1167,91 @@ def default_nodal_inj_start_options(static_information: StaticInformation):
     return NodalInjStartOptions(
         previous_results=NodalInjOptimResults(pst_taps=jnp.zeros((batch_size, n_timesteps, n_pst), dtype=jnp.float32)),
         precision_percent=jnp.array(1.0),
+    )
+
+
+def create_complex_grid_battery_hvdc_svc_3w_trafo_data_folder(folder: Path) -> None:
+    """Create a preprocessed folder for create_complex_grid_battery_hvdc_svc_3w_trafo().
+
+    Runs the importer and preprocessing.
+
+    Parameter:
+    folder: Path
+        The root folder where the data is saved to.
+    """
+    net = create_complex_grid_battery_hvdc_svc_3w_trafo()
+    pypowsybl.loadflow.run_dc(net, DISTRIBUTED_SLACK)
+
+    output_path_grid = folder / PREPROCESSING_PATHS["grid_file_path_powsybl"]
+    output_path_grid.parent.mkdir(parents=True, exist_ok=True)
+    net.save(output_path_grid)
+    output_path_masks = folder / PREPROCESSING_PATHS["masks_path"]
+    output_path_masks.mkdir(parents=True, exist_ok=True)
+
+    importer_parameters = CgmesImporterParameters(
+        grid_model_file=output_path_grid,
+        data_folder=folder,
+        area_settings=AreaSettings(
+            cutoff_voltage=1,
+            control_area=[""],
+            view_area=[""],
+            nminus1_area=[""],
+            cross_border_limits_n0=None,
+            cross_border_limits_n1=None,
+        ),
+    )
+    preprocessing_parameters = PreprocessParameters(action_set_clip=2**4, enable_bb_outage=False, bb_outage_as_nminus1=False)
+
+    _import_result = preprocessing.convert_file(importer_parameters=importer_parameters)
+    filesystem_dir = DirFileSystem(str(folder))
+    _info, _static_information, _ = load_grid(
+        data_folder_dirfs=filesystem_dir,
+        pandapower=False,
+        status_update_fn=None,
+        parameters=preprocessing_parameters,
+    )
+
+
+def create_ucte_data_folder(folder: Path, ucte_file: Path) -> None:
+    """Create a preprocessed folder for an ucte file.
+
+    Runs the importer and preprocessing.
+
+    Parameter:
+    folder: Path
+        The root folder where the data is saved to.
+    ucte_file: Path
+        The path to the UCTE file to load.
+    """
+    net = pypowsybl.network.load(ucte_file)
+    pypowsybl.loadflow.run_dc(net, DISTRIBUTED_SLACK)
+
+    output_path_grid = folder / PREPROCESSING_PATHS["grid_file_path_powsybl"]
+    output_path_grid.parent.mkdir(parents=True, exist_ok=True)
+    net.save(output_path_grid)
+    output_path_masks = folder / PREPROCESSING_PATHS["masks_path"]
+    output_path_masks.mkdir(parents=True, exist_ok=True)
+
+    importer_parameters = UcteImporterParameters(
+        grid_model_file=output_path_grid,
+        data_folder=folder,
+        area_settings=AreaSettings(
+            cutoff_voltage=1,
+            control_area=[""],
+            view_area=[""],
+            nminus1_area=[""],
+            cross_border_limits_n0=None,
+            cross_border_limits_n1=None,
+        ),
+    )
+    preprocessing_parameters = PreprocessParameters(action_set_clip=2**4, enable_bb_outage=False, bb_outage_as_nminus1=False)
+
+    _import_result = preprocessing.convert_file(importer_parameters=importer_parameters)
+
+    filesystem_dir = DirFileSystem(str(folder))
+    _info, _static_information, _ = load_grid(
+        data_folder_dirfs=filesystem_dir,
+        pandapower=False,
+        status_update_fn=None,
+        parameters=preprocessing_parameters,
     )
