--- packages/interfaces_pkg/pytest.ini
@@ -1,2 +1,0 @@
-[pytest]
-timeout_func_only = True
